Kinetis EMAC: Renamed K64F files to Kinetis
This driver is used by more SoC's than K64F
